requirements.txt
alabaster==0.7.10
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
buzio==1.1.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
CommonMark==0.5.4
configparser==3.5.0
contextlib2==0.5.5
coverage==4.0.3
decorator==4.2.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
idna==2.6
imagesize==0.7.1
ipdb==0.10.3
ipdbplugin==1.5.0
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
nose==1.3.7
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
pydocstyle==2.1.1
Pygments==2.2.0
python-coveralls==2.9.1
pytz==2017.3
PyYAML==3.12
raven==6.5.0
recommonmark==0.4.0
requests==2.18.4
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.6
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
traitlets==4.3.2
typing==3.6.4
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
